+++ src/clients/OdometryClient.js
@@ -54,7 +54,13 @@ ROS2D.OdometryClient = function(options) {
   this.rosTopic = new ROSLIB.Topic({
     ros: ros,
     name: this.topicName,
-    messageType: 'nav_msgs/Odometry'
+    messageType: 'nav_msgs/Odometry',
+    throttle_rate: options.throttle_rate,
+    queue_size: options.queue_size,
+    queue_length: options.queue_length,
+    compression: options.compression,
+    latch: options.latch,
+    reconnect_on_close: options.reconnect_on_close
   });
 
   this.rosTopic.subscribe(function(message) {