+++ src/clients/PathClient.js
@@ -44,7 +44,13 @@ ROS2D.PathClient = function(options) {
   this.rosTopic = new ROSLIB.Topic({
     ros: ros,
     name: this.topicName,
-    messageType: 'nav_msgs/Path'
+    messageType: 'nav_msgs/Path',
+    throttle_rate: options.throttle_rate,
+    queue_size: options.queue_size,
+    queue_length: options.queue_length,
+    compression: options.compression,
+    latch: options.latch,
+    reconnect_on_close: options.reconnect_on_close
   });
 
   this.rosTopic.subscribe(function(message) {