+++ src/clients/PoseStampedClient.js
@@ -62,7 +62,13 @@ ROS2D.PoseStampedClient = function(options) {
   this.rosTopic = new ROSLIB.Topic({
     ros: ros,
     name: this.topicName,
-    messageType: 'geometry_msgs/PoseStamped'
+    messageType: 'geometry_msgs/PoseStamped',
+    throttle_rate: options.throttle_rate,
+    queue_size: options.queue_size,
+    queue_length: options.queue_length,
+    compression: options.compression,
+    latch: options.latch,
+    reconnect_on_close: options.reconnect_on_close
   });
 
   this.rosTopic.subscribe(function(message) {