+++ src/maps/OccupancyGridClient.js
@@ -52,8 +52,13 @@ ROS2D.OccupancyGridClient = function(options) {
   this.rosTopic = new ROSLIB.Topic({
     ros : ros,
     name : topic,
-    messageType : 'nav_msgs/OccupancyGrid'
-    // compression : 'png'
+    messageType : 'nav_msgs/OccupancyGrid',
+    throttle_rate: options.throttle_rate,
+    queue_size: options.queue_size,
+    queue_length: options.queue_length,
+    compression: options.compression,
+    latch: options.latch,
+    reconnect_on_close: options.reconnect_on_close
   });
 
   this.rosTopic.subscribe(function(message) {